Converts signature validation middleware to typescript
